Put header components to a separate directory
