+++ site/frontend/src/pages/compare/page.vue
@@ -8,10 +8,10 @@ import {postMsgpack} from "../../utils/requests";
 import {COMPARE_DATA_URL} from "../../urls";
 import {CompareResponse, CompareSelector, DataFilter} from "./types";
 import BootstrapTable from "./bootstrap-table.vue";
-import Header from "./header.vue";
-import DataSelector, {SelectionParams} from "./data-selector.vue";
-import QuickLinks from "./quick-links.vue";
-import Filters from "./filters.vue";
+import Header from "./header/header.vue";
+import DataSelector, {SelectionParams} from "./header/data-selector.vue";
+import QuickLinks from "./header/quick-links.vue";
+import Filters from "./header/filters.vue";
 import {exportToMarkdown} from "./export";
 import {computeBenchmarkMap, computeSummary, computeTestCases} from "./data";
 import OverallTable from "./summary/overall-table.vue";
